requirements.txt
PyQt6
python-mpv
requests
beautifulsoup4
aiohttp
pyinstaller
